# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: NextSmartShip openapi
  description: NextSmartShip openapi for oms and wms
  version: ""
host: 0.0.0.0:80
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /barcode:
    post:
      tags:
      - product
      summary: generate_barcode product
      operationId: product#generate_barcode
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ProductGenerateBarcodeResponseBody'
            required:
            - barcode
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/ProductGenerateBarcodeUnauthorizedResponseBody'
      schemes:
      - http
...
  /products:
    put:
      tags:
      - product
      summary: update_product product
      operationId: product#update_product
      parameters:
      - name: update_product_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/ProductUpdateProductRequestBody'
          required:
          - product_sku
          - product_name
          - declared_en_name
          - declared_cn_name
          - declared_value_in_usd
          - product_weight
          - product_length
          - product_width
          - product_height
          - hs_code
          - product_barcode
          - product_attributes
          - customer_code
          - declared_value_in_eur
          - barcode_service
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ProductUpdateProductResponseBody'
            required:
            - status
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/ProductUpdateProductUnauthorizedResponseBody'
      schemes:
      - http
    post:
      tags:
      - product
      summary: batches_create_product product
      operationId: product#batches_create_product
      parameters:
      - name: batches_create_product_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/ProductBatchesCreateProductRequestBody'
          required:
          - products
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ProductBatchesCreateProductResponseBody'
            required:
            - results
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/ProductBatchesCreateProductUnauthorizedResponseBody'
      schemes:
      - http
...
  /token:
    post:
      tags:
      - product
      summary: generate_token product
      operationId: product#generate_token
      parameters:
      - name: generate_token_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/ProductGenerateTokenRequestBody'
          required:
          - id
          - tenant_id
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ProductGenerateTokenResponseBody'
            required:
            - token
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/ProductGenerateTokenUnauthorizedResponseBody'
      schemes:
      - http
...
definitions:
...
  MultiProductDataResponseBody:
    title: MultiProductDataResponseBody
    type: object
    properties:
      barcode:
        type: string
        description: barcode
        example: Neque asperiores.
      error_msg:
        type: string
        description: error msg
        example: Quis optio aut et.
      product_id:
        type: integer
        description: product id
        example: 6417061173984060853
        format: int64
      product_name:
        type: string
        description: product_name
        example: Repudiandae voluptates.
      sku:
        type: string
        description: sku
        example: Deleniti unde harum.
      status:
        type: string
        description: status
        example: Expedita architecto totam accusamus assumenda provident quasi.
    description: MultiProductRsp describes the multi product info
    example:
      barcode: Dicta fuga est totam qui blanditiis voluptatem.
      error_msg: Et eveniet tempore cum.
      product_id: 5199644016758294711
      product_name: At qui placeat quibusdam id eveniet.
      sku: Omnis blanditiis voluptates neque.
      status: Vel eos doloribus cum.
    required:
    - product_id
    - status
    - error_msg
    - barcode
    - sku
    - product_name
...
  ProductBatchesCreateProductRequestBody:
    title: ProductBatchesCreateProductRequestBody
    type: object
    properties:
      products:
        type: array
        items:
          $ref: '#/definitions/ProductRequestBody'
        description: product info
        example:
        - barcode_service: false
          customer_code: xxx
          declared_cn_name: xxx
          declared_en_name: xxx
          declared_value_in_eur: 50
          declared_value_in_usd: 50
          enabled_nss_barcode: false
          hs_code: xxx
          id: 1
          product_attributes:
          - liquid
          - battery
          - cosmetic
          - magnetic
          product_barcode: xxx
          product_height: 2
          product_image:
          - url
          product_length: 1
          product_name: xxx
          product_sku: xxx
          product_weight: 10
          product_width: 1
          qty: 1
        - barcode_service: false
          customer_code: xxx
          declared_cn_name: xxx
          declared_en_name: xxx
          declared_value_in_eur: 50
          declared_value_in_usd: 50
          enabled_nss_barcode: false
          hs_code: xxx
          id: 1
          product_attributes:
          - liquid
          - battery
          - cosmetic
          - magnetic
          product_barcode: xxx
          product_height: 2
          product_image:
          - url
          product_length: 1
          product_name: xxx
          product_sku: xxx
          product_weight: 10
          product_width: 1
          qty: 1
        - barcode_service: false
          customer_code: xxx
          declared_cn_name: xxx
          declared_en_name: xxx
          declared_value_in_eur: 50
          declared_value_in_usd: 50
          enabled_nss_barcode: false
          hs_code: xxx
          id: 1
          product_attributes:
          - liquid
          - battery
          - cosmetic
          - magnetic
          product_barcode: xxx
          product_height: 2
          product_image:
          - url
          product_length: 1
          product_name: xxx
          product_sku: xxx
          product_weight: 10
          product_width: 1
          qty: 1
        minItems: 1
    example:
      products:
      - barcode_service: false
        customer_code: xxx
        declared_cn_name: xxx
        declared_en_name: xxx
        declared_value_in_eur: 50
        declared_value_in_usd: 50
        enabled_nss_barcode: false
        hs_code: xxx
        id: 1
        product_attributes:
        - liquid
        - battery
        - cosmetic
        - magnetic
        product_barcode: xxx
        product_height: 2
        product_image:
        - url
        product_length: 1
        product_name: xxx
        product_sku: xxx
        product_weight: 10
        product_width: 1
        qty: 1
      - barcode_service: false
        customer_code: xxx
        declared_cn_name: xxx
        declared_en_name: xxx
        declared_value_in_eur: 50
        declared_value_in_usd: 50
        enabled_nss_barcode: false
        hs_code: xxx
        id: 1
        product_attributes:
        - liquid
        - battery
        - cosmetic
        - magnetic
        product_barcode: xxx
        product_height: 2
        product_image:
        - url
        product_length: 1
        product_name: xxx
        product_sku: xxx
        product_weight: 10
        product_width: 1
        qty: 1
    required:
    - products
  ProductBatchesCreateProductResponseBody:
    title: ProductBatchesCreateProductResponseBody
    type: object
    properties:
      results:
        type: array
        items:
          $ref: '#/definitions/MultiProductDataResponseBody'
        description: results
        example:
        - barcode: Aut doloremque aut saepe.
          error_msg: Et exercitationem est.
          product_id: 6788594783383084410
          product_name: Ut inventore architecto laboriosam architecto aut perferendis.
          sku: Excepturi adipisci optio asperiores autem enim quaerat.
          status: Quibusdam consequatur.
        - barcode: Aut doloremque aut saepe.
          error_msg: Et exercitationem est.
          product_id: 6788594783383084410
          product_name: Ut inventore architecto laboriosam architecto aut perferendis.
          sku: Excepturi adipisci optio asperiores autem enim quaerat.
          status: Quibusdam consequatur.
        - barcode: Aut doloremque aut saepe.
          error_msg: Et exercitationem est.
          product_id: 6788594783383084410
          product_name: Ut inventore architecto laboriosam architecto aut perferendis.
          sku: Excepturi adipisci optio asperiores autem enim quaerat.
          status: Quibusdam consequatur.
        - barcode: Aut doloremque aut saepe.
          error_msg: Et exercitationem est.
          product_id: 6788594783383084410
          product_name: Ut inventore architecto laboriosam architecto aut perferendis.
          sku: Excepturi adipisci optio asperiores autem enim quaerat.
          status: Quibusdam consequatur.
    example:
      results:
      - barcode: Aut doloremque aut saepe.
        error_msg: Et exercitationem est.
        product_id: 6788594783383084410
        product_name: Ut inventore architecto laboriosam architecto aut perferendis.
        sku: Excepturi adipisci optio asperiores autem enim quaerat.
        status: Quibusdam consequatur.
      - barcode: Aut doloremque aut saepe.
        error_msg: Et exercitationem est.
        product_id: 6788594783383084410
        product_name: Ut inventore architecto laboriosam architecto aut perferendis.
        sku: Excepturi adipisci optio asperiores autem enim quaerat.
        status: Quibusdam consequatur.
    required:
    - results
  ProductBatchesCreateProductUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: batches_create_product_Unauthorized_response_body result type (default
      view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ProductGenerateBarcodeResponseBody:
    title: ProductGenerateBarcodeResponseBody
    type: object
    properties:
      barcode:
        type: string
        description: barcode
        example: Dolor qui illo iusto illo nulla.
    example:
      barcode: Explicabo quisquam ea dolorem nemo.
    required:
    - barcode
  ProductGenerateBarcodeUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: generate_barcode_Unauthorized_response_body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ProductGenerateTokenRequestBody:
    title: ProductGenerateTokenRequestBody
    type: object
    properties:
      id:
        type: integer
        description: user_id
        example: 1
        minimum: 1
      tenant_id:
        type: integer
        description: tenantID
        example: 1
        minimum: 1
    example:
      id: 1
      tenant_id: 1
    required:
    - id
    - tenant_id
  ProductGenerateTokenResponseBody:
    title: ProductGenerateTokenResponseBody
    type: object
    properties:
      token:
        type: string
        description: token
        example: Enim assumenda libero et.
    example:
      token: Occaecati quisquam aliquid quidem labore sed dolores.
    required:
    - token
  ProductGenerateTokenUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: generate_token_Unauthorized_response_body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ProductRequestBody:
    title: ProductRequestBody
    type: object
    properties:
      barcode_service:
        type: boolean
        description: barcode_service
        example: false
        enum:
        - false
        - true
      customer_code:
        type: string
        description: customer code
        example: xxx
        maxLength: 50
      declared_cn_name:
        type: string
        description: declared cn name
        example: xxx
        maxLength: 50
      declared_en_name:
        type: string
        description: declared en name
        example: xxx
        maxLength: 50
      declared_value_in_eur:
        type: number
        description: declared value in eur（€）
        example: 50
        minimum: 0
      declared_value_in_usd:
        type: number
        description: declared value in usd（$）
        example: 50
        minimum: 0
      enabled_nss_barcode:
        type: boolean
        description: need use nss barcode
        example: false
        enum:
        - false
        - true
      hs_code:
        type: string
        description: hs code
        example: xxx
        maxLength: 50
      id:
        type: integer
        description: id
        example: 1
        minimum: 1
      product_attributes:
        type: array
        items:
          type: string
          example: Maiores incidunt error dolorem vel.
        description: product attributes
        example:
        - liquid
        - battery
        - cosmetic
        - magnetic
      product_barcode:
        type: string
        description: product_barcode
        example: xxx
        maxLength: 50
      product_height:
        type: number
        description: product height（mm）
        example: 2
        format: double
      product_image:
        type: array
        items:
          type: string
          example: Reprehenderit voluptatem quis quas est dolorem.
        description: product image
        example:
        - url
      product_length:
        type: number
        description: product length（mm）
        example: 1
        format: double
      product_name:
        type: string
        description: product name
        example: xxx
        maxLength: 50
      product_sku:
        type: string
        description: product sku
        example: xxx
        maxLength: 50
      product_weight:
        type: number
        description: product weight（g）
        example: 10
        format: double
      product_width:
        type: number
        description: product width（mm）
        example: 1
        format: double
      qty:
        type: integer
        description: quality
        example: 1
        format: int64
    description: Product describes the create product
    example:
      barcode_service: false
      customer_code: xxx
      declared_cn_name: xxx
      declared_en_name: xxx
      declared_value_in_eur: 50
      declared_value_in_usd: 50
      enabled_nss_barcode: false
      hs_code: xxx
      id: 1
      product_attributes:
      - liquid
      - battery
      - cosmetic
      - magnetic
      product_barcode: xxx
      product_height: 2
      product_image:
      - url
      product_length: 1
      product_name: xxx
      product_sku: xxx
      product_weight: 10
      product_width: 1
      qty: 1
    required:
    - product_sku
    - product_name
    - declared_en_name
    - declared_cn_name
    - declared_value_in_usd
    - product_weight
    - product_length
    - product_width
    - product_height
    - hs_code
    - product_barcode
    - product_attributes
    - customer_code
    - declared_value_in_eur
    - barcode_service
  ProductUpdateProductRequestBody:
    title: ProductUpdateProductRequestBody
    type: object
    properties:
      barcode_service:
        type: boolean
        description: barcode_service
        example: false
        enum:
        - false
        - true
      customer_code:
        type: string
        description: customer code
        example: xxx
        maxLength: 50
      declared_cn_name:
        type: string
        description: declared cn name
        example: xxx
        maxLength: 50
      declared_en_name:
        type: string
        description: declared en name
        example: xxx
        maxLength: 50
      declared_value_in_eur:
        type: number
        description: declared value in eur（€）
        example: 50
        minimum: 0
      declared_value_in_usd:
        type: number
        description: declared value in usd（$）
        example: 50
        minimum: 0
      enabled_nss_barcode:
        type: boolean
        description: need use nss barcode
        example: false
        enum:
        - false
        - true
      hs_code:
        type: string
        description: hs code
        example: xxx
        maxLength: 50
      id:
        type: integer
        description: id
        example: 1
        minimum: 1
      product_attributes:
        type: array
        items:
          type: string
          example: Dolor nihil ipsum consequatur et voluptatem asperiores.
        description: product attributes
        example:
        - liquid
        - battery
        - cosmetic
        - magnetic
      product_barcode:
        type: string
        description: product_barcode
        example: xxx
        maxLength: 50
      product_height:
        type: number
        description: product height（mm）
        example: 2
        format: double
      product_image:
        type: array
        items:
          type: string
          example: Mollitia labore quia.
        description: product image
        example:
        - url
      product_length:
        type: number
        description: product length（mm）
        example: 1
        format: double
      product_name:
        type: string
        description: product name
        example: xxx
        maxLength: 50
      product_sku:
        type: string
        description: product sku
        example: xxx
        maxLength: 50
      product_weight:
        type: number
        description: product weight（g）
        example: 10
        format: double
      product_width:
        type: number
        description: product width（mm）
        example: 1
        format: double
      qty:
        type: integer
        description: quality
        example: 1
        format: int64
    example:
      barcode_service: false
      customer_code: xxx
      declared_cn_name: xxx
      declared_en_name: xxx
      declared_value_in_eur: 50
      declared_value_in_usd: 50
      enabled_nss_barcode: false
      hs_code: xxx
      id: 1
      product_attributes:
      - liquid
      - battery
      - cosmetic
      - magnetic
      product_barcode: xxx
      product_height: 2
      product_image:
      - url
      product_length: 1
      product_name: xxx
      product_sku: xxx
      product_weight: 10
      product_width: 1
      qty: 1
    required:
    - product_sku
    - product_name
    - declared_en_name
    - declared_cn_name
    - declared_value_in_usd
    - product_weight
    - product_length
    - product_width
    - product_height
    - hs_code
    - product_barcode
    - product_attributes
    - customer_code
    - declared_value_in_eur
    - barcode_service
  ProductUpdateProductResponseBody:
    title: ProductUpdateProductResponseBody
    type: object
    properties:
      status:
        type: integer
        description: status
        example: 1
        enum:
        - 0
        - 1
    example:
      status: 1
    required:
    - status
  ProductUpdateProductUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: update_product_Unauthorized_response_body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...